Switching the flute instrument aftertouch message to (int, int) rather than (int, double). Ultimately it might be prudent for all noteOn NoteOff and aftertouch messages from the Skoog world to conform to the MIDI spec and they should probably only be float or double in the Skoog engine when first decoded and again at the end of the chain when STK::Flute for example accepts control changes on various channels in the range of 0-127 and then scales them back to 0...1 for use in the STK world. We can of course convert back and forth at any time to suit the instrument in question, but it would be neater if all of them behaved the same way and we could confidently send them 0-127 to any of the three main functions (notion, noteoff, aftertouch) in any of our instruments and get expected results.

+++ AudioKit/Common/Nodes/Generators/Physical Models/AKFluteInstrumentAudioUnit.h
@@ -16,7 +16,7 @@
 @property (nonatomic) float amplitude;
 
 - (void)triggerFrequency:(float)frequency amplitude:(float)amplitude;
-- (void)controlChange:(int)channel value:(float)value;
+- (void)controlChange:(int)channel value:(int)value;
 - (void)startNote:(int)note velocity:(int)vel;
 - (void)stopNote:(int)note;
 - (void)start;
